Properly dispose all disposable member objects
Objects obtained through DI are never disposed, because the DI container handles that for us due to how we register them to it.

All implementations of IDisposable and/or IAsyncDisposable will never attempt to dispose objects twice. The way this is achieved differs between implementations (primarily if IAsyncDisposable also has to be implemented).

Notably, in the DiscordManager implementation of IAsyncDisposable, we do *not* check if the Discord client is disposable or asynchronously disposable, as this will always be true due to the client being a concrete type SocketDiscordClient, which implements both interfaces already.